Allow mod creators to use info.toml to specify installation paths
If there is an info.toml file at the root of the gamebanana archive, and it has a "paths" attribute, we read that in. The paths attribute should look like

paths = [
    ["atmosphere", ""],
    ["MyModFolder", "ultimate/mods/"],
]
you put the source file/folder you want to move, then the destination it should be moved to on the user's switch.
Empty strings -> sd card root